Files renamed according to classes used or intended to use
