+++ ast_explore_lite/CMakeLists.txt
@@ -49,9 +49,10 @@ include_directories(
 )
 
 add_executable(ast_explore_lite
-  src/costmap_client.cpp
-  src/explore.cpp
-  src/frontier_search.cpp)
+        src/Costmap2DClient.cpp
+        src/Explore.cpp
+        src/FrontierSearch.cpp
+        src/Frontier.cpp)
 add_dependencies(ast_explore_lite ${${PROJECT_NAME}_EXPORTED_TARGETS} ${catkin_EXPORTED_TARGETS})
 target_link_libraries(ast_explore_lite ${catkin_LIBRARIES})
