+++ ast_explore_lite/include/explore/Explore.h
@@ -48,8 +48,8 @@
 #include <ros/ros.h>
 #include <visualization_msgs/MarkerArray.h>
 
-#include <explore/costmap_client.h>
-#include <explore/frontier_search.h>
+#include <explore/Costmap2DClient.h>
+#include <explore/FrontierSearch.h>
 
 namespace explore
 {
